pyproject.toml
[tool.poetry]
name = "web_2021_django_react_passwordmnager"
version = "0.1.0"
description = ""
authors = ["pushsla <push2001sla@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
cryptography = "^36.0.0"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
